Added fundation for real i18n support (JSP only)
* Removed the jsp-api dependency from the core artefact
* New LocaleResolver and MessageResolver interfaces and their default
implementation
* New JstlLocaleResolver and JstlMessageResolver implementations (to be
able to use the new titleKey attribute of the ColumnTag as fmt:message)
* New SpringLocaleResolver implementations
* New way to load user properties, now using ResourceBundle. The
ResourceBundle can be loaded either from the classpath or from any
filesystem using by setting a system property
* The default properties file now contains new key for all
internationalized messages
* New DatatablesMsg enum
* New MessageProcessor, allowing to update the TableConfiguration
instance with the user-defined messages
* All prototypes of TableConfiguration are now stored in the
ConfigurationStore class. A prototype is created for each group and for
each locale.
* Unit tests updated accordingly
* Linked with dandelion/issues#4